GPU OpenCL: Reorganize OpenCL code and remove obsolete internals header
