+++ GPU/GPUTracking/Base/opencl/GPUReconstructionOCL.h
@@ -59,8 +59,8 @@ class GPUReconstructionOCLBackend : public GPUReconstructionDeviceBase
   int32_t AddKernel(bool multi = false);
   template <class T, int32_t I = 0>
   uint32_t FindKernel(int32_t num);
-  template <typename K, typename... Args>
-  void runKernelBackendInternal(const krnlSetupTime& _xyz, K& k, const Args&... args);
+  template <class T, int32_t I = 0, typename... Args>
+  void runKernelBackendInternal(const krnlSetupTime& _xyz, const Args&... args);
   template <class T, int32_t I = 0>
   gpu_reconstruction_kernels::krnlProperties getKernelPropertiesBackend();
